% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apartments.R
\name{adj_ppsize}
\alias{adj_ppsize}
\title{Function to adjust price by unit size}
\usage{
adj_ppsize(df, SubjAvgUnitSF)
}
\arguments{
\item{df}{data frame containing AvgUnitSF, PPUnit, AdjPPUnit (from adj_ppunit function)}

\item{SubjAvgUnitSF}{the average square ft of the subject property}
}
\value{
returns a data frame with additional column: IndPPUnit
}
\description{
Function to adjust price by unit size
}
