% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apartments.R
\name{adj_ppunit}
\alias{adj_ppunit}
\title{Function to calculate adjusted price per unit}
\usage{
adj_ppunit(df, from_date, value_date)
}
\arguments{
\item{df}{data frame with Date Sold and PPUnit}

\item{from_date}{date to calculate adjustment factor from}

\item{value_date}{date of valuation}
}
\value{
returns a data frame with additional column: AdjPPUnit
}
\description{
Function to calculate adjusted price per unit
}
