% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apartments.R
\name{avg_unitsf}
\alias{avg_unitsf}
\title{Calculates average unit square foot}
\usage{
avg_unitsf(df)
}
\arguments{
\item{df}{data frame with number of units (NoUnits) and square foot (SqFt)}
}
\value{
returns a data frame with additional column: AvgUnitSF
}
\description{
Calculates average unit square foot
}
