% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{costar_property}
\alias{costar_property}
\title{Title}
\usage{
costar_property(input_file_path, output_file)
}
\arguments{
\item{input_file_path}{file path where costar xlsx files are kept}

\item{output_file}{output file path and filename}
}
\value{
reads in data, cleans and saves to rds file
}
\description{
Title
}
