% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{define_subject}
\alias{define_subject}
\title{Define attributes of subject property}
\usage{
define_subject(subject_lat, subject_lng)
}
\arguments{
\item{subject_lat}{The subject properties latitude}

\item{subject_lng}{The subject properties longitude}
}
\value{
assigns the attributes to the global environment
}
\description{
Define attributes of subject property
}
