% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{facet_boxplot}
\alias{facet_boxplot}
\title{Facet boxplot}
\usage{
facet_boxplot(df, variable)
}
\arguments{
\item{df}{data frame}

\item{variable}{variable to examine across the different clusters}
}
\value{
returns boxplots of the named variable for all the different clusters
}
\description{
Facet boxplot
}
