% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{facet_hist}
\alias{facet_hist}
\title{Faceted histogram}
\usage{
facet_hist(df, variable)
}
\arguments{
\item{df}{data frame}

\item{variable}{variable to examine across the different clusters}
}
\value{
returns histograms of the named variable for all the different clusters
}
\description{
Faceted histogram
}
