% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cknn.R
\name{get_nn}
\alias{get_nn}
\title{Return nearest neighbours data}
\usage{
get_nn(df, subject_pid = NULL)
}
\arguments{
\item{df}{data frame with nearest neighbours information}

\item{subject_pid}{subject_pid - default = NULL for Rmd}
}
\value{
returns data frame with just nearest neighbours and subject properties data
}
\description{
Return nearest neighbours data
}
