% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cknn.R
\name{get_subj_cluster}
\alias{get_subj_cluster}
\title{return subject cluster}
\usage{
get_subj_cluster(df)
}
\arguments{
\item{df}{data frame with subject property and cluster information}
}
\value{
returns the subject cluster
}
\description{
return subject cluster
}
