% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{leaflet_plot}
\alias{leaflet_plot}
\title{Create a leaflet plot of the subject and comparable properties}
\usage{
leaflet_plot(df, subject_lng, subject_lat)
}
\arguments{
\item{df}{Data frame with comparable properties}

\item{subject_lng}{subject properties longitude}

\item{subject_lat}{subject properties latitude}
}
\value{
returns a leaflet plot with subject and comparable properties
}
\description{
Create a leaflet plot of the subject and comparable properties
}
\examples{
Modified from g2- Case Study.Rmd:
leaflet_plot(Apartments2[1:260,], subject_lng = -117.141637, subject_lat = 32.756933)
}
