% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apartments.R
\name{no_br}
\alias{no_br}
\title{Calculating total number of bedrooms}
\usage{
no_br(df)
}
\arguments{
\item{df}{Data frame containing info on number of bedrooms}
}
\value{
returns a data frame with additional column of NoBR
}
\description{
Calculating total number of bedrooms
}
