% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cknn.R
\name{optimal_m}
\alias{optimal_m}
\title{Finding the optimum number of clusters}
\usage{
optimal_m(df, m_values = 3:15, n_start = 5)
}
\arguments{
\item{df}{data frame}

\item{m_values}{m (number of clusters) values to test}

\item{n_start}{number of repetitive computations of kproto with random initializations}
}
\value{
returns optimum number of clusters according to the silhouette index.
}
\description{
Finding the optimum number of clusters
}
