% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_knn}
\alias{plot_knn}
\title{Plots subject property with k nearest neighbours from cluster}
\usage{
plot_knn(df)
}
\arguments{
\item{df}{data frame from recombine_data_knn function}
}
\value{
returns leaflet plot, with subject property and k nearest neighbours
}
\description{
Plots subject property with k nearest neighbours from cluster
}
