% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{price_avgUnitSF_scatter}
\alias{price_avgUnitSF_scatter}
\title{Title}
\usage{
price_avgUnitSF_scatter(df, ylimits = c(0, 350000))
}
\arguments{
\item{df}{data frame}

\item{ylimits}{limits of y axis}
}
\value{
returns a scatter plot of reported sale price per unit vs reported average unit size,
including a fitted line of best fit.
}
\description{
Title
}
\examples{
Modified from g2- Case Study.Rmd:
price_avgUnitSF_scatter(NPApts)
}
