% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{price_boxplot}
\alias{price_boxplot}
\title{Create a boxplot of different sales metrics}
\usage{
price_boxplot(
  df,
  variable = c("PriceSold", "PPSF", "PPBR", "PPUnit"),
  ylimits = c(0, 7500000)
)
}
\arguments{
\item{df}{Data frame}

\item{variable}{variable of interest: PriceSold, PPSF, PPBR or PPUnit}

\item{ylimits}{limits of y axis}
}
\value{
returns a boxplot
}
\description{
Create a boxplot of different sales metrics
}
\examples{
Modified from g2- Case Study.Rmd:
price_boxplot(Apartments3, variable = 'PPSF', ylimits = c(0,750))
}
