% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apartments.R
\name{price_per}
\alias{price_per}
\title{Creates price per metrics based off of price sold}
\usage{
price_per(df)
}
\arguments{
\item{df}{Data frame with Price Sold information}
}
\value{
returns a data frame with additional columns: PPSF, PPUnit and PPBR
}
\description{
Creates price per metrics based off of price sold
}
