% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cknn.R
\name{recombine_data_knn}
\alias{recombine_data_knn}
\title{recombines knn output with original data}
\usage{
recombine_data_knn(initial_data, knn_object)
}
\arguments{
\item{initial_data}{original data frame}

\item{knn_object}{knn object from cknn function}
}
\value{
returns original data frame with knn information
}
\description{
recombines knn output with original data
}
