% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reproject_latlon}
\alias{reproject_latlon}
\title{Reproject lat/lon into planar coordinates}
\usage{
reproject_latlon(df, old_crs = 4326, new_crs = 3502)
}
\arguments{
\item{df}{data frame containing Latitude and Longitude}
}
\value{
returns data frame with Latitude and Longitude reprojected into EPSG 3502 (central colorado)
}
\description{
Reproject lat/lon into planar coordinates
}
\examples{
reproject_latlon(sales_with_knn, old_crs = 3502, new_crs = 4326)
}
