% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sales_by_month}
\alias{sales_by_month}
\title{Average and Median Sales Price by Month}
\usage{
sales_by_month(df)
}
\arguments{
\item{df}{Data frame with date_sold and price_sold variables}
}
\value{
returns plot of Average and Median Sales Price by Month
}
\description{
Average and Median Sales Price by Month
}
\examples{
sales_by_month(Apartments2)
}
