% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sales_time_scatter}
\alias{sales_time_scatter}
\title{Title}
\usage{
sales_time_scatter(df, ylimits = c(0, 350000))
}
\arguments{
\item{df}{data frame}

\item{ylimits}{limits of y axis}
}
\value{
returns a scatter plot of reported sale price per unit vs reported sale data,
including a fitted line of best fit.
}
\description{
Title
}
\examples{
Modified from g2- Case Study.Rmd:
sales_time_scatter(Apartments4)
}
