% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{standardise_colnames}
\alias{standardise_colnames}
\title{Standardises the variables names from different datasets.}
\usage{
standardise_colnames(df, data_src, data_mapping)
}
\arguments{
\item{df}{dataframe where you want to rename the variables}

\item{data_src}{data source type (e.g. 'valcre', 'epc')}

\item{data_mapping}{data mapping table - included in bearing package - data('data_mapping')}
}
\value{
returns dataframe with renamed variables
}
\description{
Standardises the variables names from different datasets.
}
