% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{define_subject}
\alias{define_subject}
\title{Define attributes of subject property}
\usage{
define_subject(df, subject_pid)
}
\arguments{
\item{df}{data frame containing subject property}

\item{subject_pid}{ID of subject property}
}
\value{
assigns the attributes to the global environment
}
\description{
Define attributes of subject property
}
