% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cknn.R
\name{get_nn}
\alias{get_nn}
\title{TReturn nearest neighbours data}
\usage{
get_nn(df)
}
\arguments{
\item{df}{data frame with nearest neighbours information}
}
\value{
returns data frame with just nearest neighbours and subject properties data
}
\description{
TReturn nearest neighbours data
}
