% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot subject property with other properties from same cluster}
\usage{
plot_clusters(df, subject_ap = subject_pid, subj_cluster = NULL)
}
\arguments{
\item{df}{data frame from recombine_data_knn function}

\item{subject_ap}{subject property id}

\item{subj_cluster}{subject property cluster}

\item{...}{}
}
\value{
returns leaflet plot, with subject property and other properties from same cluster
}
\description{
Plot subject property with other properties from same cluster
}
