% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{price_boxplot}
\alias{price_boxplot}
\title{Create a boxplot of different sales metrics}
\usage{
price_boxplot(
  df,
  variable = c("All", "PriceSold", "PPSF", "PPBR", "PPUnit"),
  columns = 1
)
}
\arguments{
\item{df}{Data frame}

\item{variable}{variable of interest: PriceSold, PPSF, PPBR or PPUnit. default = All.}

\item{ylimits}{limits of y axis}
}
\value{
returns a boxplot
}
\description{
Create a boxplot of different sales metrics
}
\examples{
Modified from g2- Case Study.Rmd:
price_boxplot(Apartments3, variable = 'PPSF')
}
