% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sales_price_histogram}
\alias{sales_price_histogram}
\title{Create a histogram of different sales metrics}
\usage{
sales_price_histogram(
  df,
  variable = c("All", "PriceSold", "PPSF", "PPBR", "PPUnit"),
  columns = 2
)
}
\arguments{
\item{df}{Data frame}

\item{variable}{variable of interest: PriceSold, PPSF, PPBR or PPUnit.}

\item{columns}{number of columns for plots - default = 2}
}
\value{
returns a histogram
}
\description{
Create a histogram of different sales metrics
}
\examples{
Modified from g2- Case Study.Rmd:
sales_price_histogram(Apartments3, variable = 'PriceSold', bin_width = 100000, xlimits = c(0,7500000))
}
