% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sales_price_histogram}
\alias{sales_price_histogram}
\title{Create a histogram of different sales metrics}
\usage{
sales_price_histogram(
  df,
  variable = c("PriceSold", "PPSF", "PPBR", "PPUnit"),
  bin_width = 1e+05,
  xlimits = c(0, 7500000)
)
}
\arguments{
\item{df}{Data frame of}

\item{variable}{variable of interest: PriceSold, PPSF, PPBR or PPUnit}

\item{bin_width}{width of histogram bins}

\item{xlimits}{x axis limits}
}
\value{
returns a histogram
}
\description{
Create a histogram of different sales metrics
}
\examples{
Modified from g2- Case Study.Rmd:
sales_price_histogram(Apartments3, variable = 'PriceSold', bin_width = 100000, xlimits = c(0,7500000))
}
