% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sales_time_scatter}
\alias{sales_time_scatter}
\title{Title}
\usage{
sales_time_scatter(df, sales_var = "price_sold", group_var = "subject_m")
}
\arguments{
\item{df}{data frame}

\item{sales_var}{sales variable to use for plot. Default = price_sold}

\item{group_var}{variable to group on. Default = subject_m}
}
\value{
returns a scatter plot of a sales price variable vs reported sale date,
including a fitted line of best fit.
}
\description{
Title
}
\examples{
sales_time_scatter(sales_with_knn, sales_var = 'price_sold')
}
